Update lfs.go
Update to clean up the use of the 0755 value in the directory creation code. Also added comment about using the system's umask value in the future to better align created locations with that of the system's admin. Refactoring here should make updates easier in the future.

NOTE: Not sure about the actual syntax of Go and its ability to keep uint32 values true given their assignment as constants.