Isolate init tests
There's a race condition in the integration suite for anything that
relies on $HOME/.gitconfig. When the suite runs in parallel (by
default), a new $HOME/.gitconfig is copied in for each test that's run.
If a test is trying to modify and validate changes to this file, as the
init tests are, then it could be overwritten before the checks, causing
tests to fail.