bootstrap: setup GOPATH if it isn't already set.
This variable is a bit of a mystery to non-Go developers and it's
unclear how you set it up properly. As a result, if it's not set let's
set everything up how Go likes it. Ignore the `src` directory that
we're creating now.

This will also help keep the Homebrew formula simpler.

Closes #388.