Update LFS license file to reflect MIT license use
Adding language at the top of the file that makes clear that the project is licensed under the MIT license, with a link to the Wikipedia article explaining what that license is. 

Users should have an easy way to determine what standard OSS license applies to a project, if any, by quickly looking at it.