Resolve symlink on GitDir to get real_path as RootDir
git.RootDir() is from command `git rev-parse --show-toplevel`, which
returns real_path (symlinks are resolved), but git.GitDir() is not.
So if git worktree is on a symlink directory, LocalWorkingDir and
LocalGitDir will be quite different.

Resolve symlinks on git.GitDir() to make it a real_path as git.RootDir().

Signed-off-by: Jiang Xin <worldhello.net@gmail.com>