Disable exit code checking on `git diff-index` to avoid breaking fsck
Problem is that the clean filter will fail earlier because of corruption
and this in turn fails the diff-index, which we now pick up. But this
then fails ScanIndex which fsck treats as fatal instead of realising it's
a symptom of something it can fix. Not sure how to handle this better yet