Fix problems with user prompts in `git lfs clone`
Stop using the psuedo-tty & pass through streams directly. Getting pty
to work with stdin was too difficult, even when it worked passwords were
displayed on the terminal which is no good. We no longer need to filter
the stderr for known warnings; Git 2.8 is now released which deals with
them and for Git 2.7 and below we can fall back on `cat` instead of a
blank filter; it's slightly slower but eliminates the problem more
simply.