git/git: sanitize pattern in git.GetTrackedFiles
`git ls-files` does not play nicely with patterns that contain a leading slash,
such as "/*.txt". To work around this, if a leading slash is present, it is
removed before being passed to the git call. Furthemore, the results are
checked such that they only match files that are direct descendants of the
repository's root, as according to:

https://git-scm.com/docs/gitignore#_pattern_format