In git 2.9+, run "git lfs pull" in submodules after "git lfs clone"
git 2.9+ preserves the config settings which disable the LFS filters in
submodules so any LFS files will be pointers. This makes it more optimal
(previous versions will suppress these settings so the smudge filter
will be used in submodules) but we must run "git lfs pull" to update
the submodule working copies.