config: implement type `EnvFetcher`
The `EnvFetcher` is an implementation of the `Fetcher` type to match the
existing behavior around interacting with environment variables as of v1.3.1.

To summarize: we have a type (safe to use across multiple goroutines) that
knows how to fetch and cache values stored via `os.Getenv`.

As per some existing behavior, a few methods that I'd like to get rid of, remain:
  * `EnvFetcher.Set(k, v string) error`
  * `EnvFetcher.SetAll(vv map[string]string)`

These methods are remnants of the mutable behavior that the `*config.Config`
type currently supports. I plan to remove these methods in favor of an
immuatable `*EnvFetcher` implementation, and simply have tests create a new
`*config.Config` when they want to change values. Not only will this make the
implementation of the `config` package much easier, but it will *greatly*
clarify the sort of behavior that is going on in the tests.