WIP: promised downloads (proof of concept)
If a file is not in the GitLFS cache, then GitLFS returns an empty file
upon Git's smudge request and starts the download right away. At the end
of the "process-filter" protocol GitLFS waits for the downloads to
finish and writes the files to the Git working tree.

This allows us to get rid of the git lfs clone|pull|fetch commands.