filter-process: fix reading 1024 byte files
If we read exactly the number of bytes that fit into `p` then
implementation detected already an "overfilled" buffer and returned
the number of read bytes `n` and `nil` as error.

Fix this by treading the exact number not as "overfill". This way the
next `readPacket()` call we will read an empty chunk and `read()` will
return the number of read bytes `n` and `o.EOF` as (legitmate) error.

c.f. https://github.com/git-lfs/git-lfs/issues/1697