Fix debian 9 (stretch) builds
This commit skips the `dh_golang` step which is used to add the
Built-using field to the resulting deb.  The field is only used to track
the version of go used to build a package so omitting it shouldn't be a
problem.  In the commit at https://anonscm.debian.org/cgit/collab-maint/dh-golang.git/commit/script/dh_golang?id=7c3fbec6ea92294477fa8910264fe9bd823f21c3
the dh-golang maintainers changed the logic used to determine the
version of go used to check the installed package.  This fails because
git-lfs gets built using a manually installed version of go.