installer: Insert the app directory at the beginning of PATH
This way binaries that come with this installer are given precedence
over any previously installed Git LFS binaries shipping with Git for
Windows. Users can then run this installer to get a more recent version
of Git LFS than the one coming with Git for Windows without the need to
wait for a new Git for Windows release with updated bundled Git LFS
binaries.

Closes #2587.