Allow tq.Manifest to depend on operation and remote
The mechanism will be used in a later commit to implement per-remote
standalone transfer agents.  The transfer queue config will then depend
on the endpoint, which depends on the operation and the remote.
Operation and remote must, therefore, be available when constructing the
`tq.Manifest`.

Passing strings "download" and "upload" feels a bit fragile.  But it is
how `EndpointFinder` expects operations.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>