commands/{un,}track: local-prefix agnostic comparison
This patch use the trimCurrentPrefix utility introduced in the previous
patch to make all .gitattributes comparisons agnostic to the
"current-prefix" of "./" or ".\".

Since we treat both sides of the comparison as having passed through
"trimCurrentPrefix", we harden ourselves against "legacy"-style
.gitattributes patterns containing prefixes, and "modern"-style ones
lacking prefixes.

That means that both combinations of "git lfs untrack ./a.dat" and "git
lfs untrack a.dat" will remove both of these lines from a .gittatributes
file:

  ./a.dat filter=lfs ...
  a.dat filter=lfs ...

Signed-off-by: Taylor Blau <me@ttaylorr.com>