Add hint that file:// URLs require a standalone transfer agent
Local remotes can only work with a standalone transfer agent.  This
commit adds a hint that refers to the documentation for details.  The
output of `git lfs fetch`, for example, now looks like:

```
fetch: Fetching reference refs/heads/master
Downloading LFS objects:   0% (0/1), 0 B | 0 B/s
hint: The remote resolves to a file:// URL, which can only work with a
hint: standalone transfer agent.  See section "Using a Custom Transfer Type
hint: without the API server" in custom-transfers.md for details.
batch request: missing protocol: "file:///..."
error: failed to fetch some objects from 'file:///...'
```

The hint is printed in `NewRequest()`, when it is clear that an actual
request is about to fail.  A potential alternative would be to print the
hint earlier, for example in `NewManifest()` or `NewEndpoint()`.  But
being too eager could cause printing spurious hints that would not be
followed by an error, which might be confusing.

Signed-off-by: Steffen Prohaska <prohaska@zib.de>