lfsapi/ssh.go: use zero-value sentinels
Instead of storing the ExpiresAt and ExpiresIn fields as a *time.Time,
and *int, respectively, use a time.Time and int instead.

Later, in the zero-value comparison, change from != nil to != 0 and
time.Time.IsZero(), respectively.

This has the effect of ridding the API of setting expires_in to zero,
and will fall-back to the default expiration offset, after which it will
retry, or fail the underlying request. We don't have precise
determinations on how many API requests this will affect, but it's
certainly not common usage of the specification, so the behavior change
is OK.