vendor: update github.com/git-lfs/wildmatch
In a subsequent commit, we will introduce the git/gitattr package, which
will emulate a tree-like structure of .gitattribute(s) files through a
repository.

In order to implement this behavior correctly, we need to retain a
behavior from Git where single-directory patterns (like '*.txt', and
'foo.bar', but not 'path/to/*.txt') match against the basename of a
filepath, not its fully-qualified path.

This behavior was introduced upstream in package
github.com/git-lfs/wildmatch via [1], and will be used here in the
following commit.

As such, let's update our dependency on wildmatch in this commit so that
we can use this feature in the following commit.

[1]: b31c344 (Merge pull request #8 from git-lfs/wm-basename,
     2018-07-06)