fs: support multiple object alternates
Like in upstream Git, Git LFS supports "object database alternates"
(described in [1]) to link several object databases together.

We have supported this feature since [2] but read the entire
.git/objects/info/alternates in one single buffer via ioutil.ReadFile
and checked if a directory existed at that location.

If a user specified multiple alternates, then this will fail. Git LFS
will look for a directory that has newlines in it, when what it should
be doing is splitting that list by the '\n' character and checking each
directory if it exists.

Let's do that, and also return a []string, instead of a single string,
to indicate that there might be multiple alternates in a repository.

[1]: https://git-scm.com/docs/gitrepository-layout#gitrepository-layout-objectsinfoalternates
[2]: https://github.com/git-lfs/git-lfs/pull/1007/files