commands/command_migrate_import.go: respect un-set filters
If a later .gitattributes sets filter to something other than "lfs", or
implicitly does so with

    *.txt !filter

We should not migrate that file. Hence, while considering only
attributes that set "filter", set "ok" to whether or not it equals
"lfs".

Because attributes are guaranteed to be considered in topological order
according to directory layout, this will consider the correct order of
precedence.