lfs/gitfilter_smudge.go: write-over locked files
This pull request teaches the deprecated `git lfs checkout` command to
be able to checkout read-only files in the working copy.

If a file tracked with Git LFS is locked, it will not have the writing
bit unless the user themselves holds the lock. In the former case, if a
user wants to check out a locked Git LFS file--provided that they will
not push a changed version of that file--they should be allowed to check
that file out.

To do so, let's recognize when a file is marked as read-only,
temporarily grant it writing permissions, and then restore the original
permissions once we're done.