commands/command_migrate.go: print newline conditionally
Sometimes when invoking 'git lfs migrate import', mysterious output like
the following can be seen:

    $ git lfs migrate import
    migrate: override changes in your working copy? [Y/n] y

    migrate: changes in your working copy will be overridden ...
    migrate: Fetching remote refs: ..., done

Where an extra newline is printed between the answer, 'y', and the next
line of output from the migrator.

Instead, let's only print that secondary newline when one isn't given in
the answer. This should never be the case (c.f., ReadString()), but will
harden the code to changes like opening /dev/tty in raw mode and reading
character-by-character.

To do so, ensure that the answer doesn't satisfy:

    strings.HasSuffix(answer, '\n')

...and print a newline iff it doesn't.

In a related sense, ignore io.EOF errors.