fs/fs.go: strings.HasPrefix typo
existsAlternate supposes a valid alternate string, and sanitizes it. One
case of this is unquoting a quoted string using strconv.Unquote, which
we trigger if and only if the string begins with a double-quote
character, ".

This function was extracted from a similar-looking function that ignores
comments (i.e., those strings beginning with '#'), so a typo was made
here.

This patch fixes that.