test/test-migrate-fixtures.sh: avoid newline in a.txt
Newlines can cause problems with the 'calc_oid()' function, which
ordinarily isn't a problem since we don't inspect the contents of a.txt
in order to calculate an OID in any of the test/test-migrate-*.sh tests.

In the following commit, this will change. So, without loss of
generality, let's change the contents of a.txt to no longer contain a
newline.