config: add origin to remote list
Previously, origin was being explicitly excluded from the list
of remotes. This could cause the wrong remote to be used in some
cases, such as with `git lfs pull`, and for the `origin` remote to
be missed in `git lfs env`. This commit ensures that the remote
list has the correct contents, and that `git lfs pull` and
`git lfs env` now function as expected.