script/install.sh: mark as executable
In [1], it is noted that script/install.sh is no longer executable when
run from the tarball'd and zip'd release artifacts that we distribute as
a part of the normal Git LFS release process.

This is due to 22351986 (Makefile: replace script/bootstrap with
'make', 'make all', 2018-07-19), where this file lost permission
100755 and gained permission 100644.

The later is no longer executable, but needs to be. So, let's restore it
as such.

[1]: https://github.com/git-lfs/git-lfs/issues/3154