t: use 'tar' instead of 'zip'
In [1], we used 'zip(1)' to create *.zip files in these integration
tests because ZIP files begin with the header:

    []byte{0x50, 0x4B, 0x03, 0x04}

And thus are detected as 'application/zip' by http.DetectContentType().

However, this 'zip(1)' program does not exist on Windows, hence the
recent test failures (starting as early as [1]).

Instead, let's use 'tar -czf', which creates gzip-compressed tarballs,
and begin with the header:

    []byte{0x1F, 0x8B, 0x80}

Thus are detected as 'application/x-gzip' by http.DetectContentType().

This should allow this test to run on AppVeyor (and more generally, on
Windows as a whole) again.

[1]: 938df5c4 (tq/basic_upload.go: disable Content-Type detection by
     configuration, 2018-07-31)