fetch: add option to exclude OIDs
In exceptional cases Git LFS objects might not be available anymore on
a Git LFS server. This leads to errors if those objects are still
referenced in a Git repository. The only way to overcome these errors
is to rewrite the Git history and remove the "dangling" pointers.
Unfortunately, a history rewrite is not be possible in every case.

Therefore, add a config "lfs.fetchexcludeoids" to exclude certain Git
LFS object IDs. The multiple object IDs are separate by a comma. The
config can be set via the local Git config mechanism or via the
.lfsconfig mechanism.

Example:
[lfs]
    fetchexcludeoids = "hash1, hash2"