Always provide a Content-Type when uploading files
As of v2.5.1, we provide an option to disable automatic detection of the
Content-Type and always use application/octet-stream instead.  However,
when that option is set, instead of providing a Content-Type of
application/octet-stream, we fail to provide one at all.

Ensure that we always set some Content-Type header when uploading files
by skipping the detection but preserving the fallback behavior when the
option is set.