debian/rules: unset GO15VENDOREXPERIMENT
This option was introduced historically in order to build Go packages in
Go 1.5 with vendored dependencies in the vendor tree.

GO15VENDOREXPERIMENT has long since been deprecated, and is no longer
required in modern versions of Go in order to instruct to the 'go' tool
to look in this directory.

Co-authored-by: brian m. carlson <bk2204@github.com>