Makefile: force callers to have goversioninfo in $PATH
In 22351986 (Makefile: replace script/bootstrap with 'make', 'make all',
2018-07-19), the Makefile learned how to make binaries of git-lfs,
including invoking 'go generate' in order to generate the resource.syso,
necessary for building on Windows.

In the commit mentioned above, we taught the Makefile recipe for
'resource.syso' to 'go get' the necessary 'goversioninfo' binary before
building, if callers do not have it.

In Go 1.10.3 and earlier, this caused no change to the vendor directory
or glide.yaml file, and instead installed the project in '$GOPATH', and
built the binary in '$GOPATH/bin'.

Since we are on Go 1.11, now, and running 'go get' from inside a module
directory causes the module's go.mod (and thusly, the go.sum, too) to
change, let's make it the caller's responsibility to provide a valid
'goversioninfo' binary in '$PATH'.

This allows us to avoid the burden of cleaning up after ourselves when
running 'go get' from within a directory that is itself a Go module.