go.mod: pin github.com/git-lfs/wildmatch to v1.0.0
In order to prefer semantically versioned tags of our dependent
repositories over Go's date & SHA-1 format, let's tag
github.com/git-lfs/wildmatch.git at v1.0.0 (on the latest master) and
vendor it as such here.

The only change between b31c34466d64 and v1.0.0 is:

  * 83d2acb (Merge pull request #9 from git-lfs/ttaylorr/go-1.11,
    2018-08-31)
  * 2f71dd1 (.travis.yml: build on Go 1.11, 2018-08-30)
  * 4bab7d7 (go.mod: initial commit, 2018-08-30)

Which is the introduction of the go.mod file (and related changes).
Thus, there are no code changes between the two, so this is a safe
change to make.