git: add a function to look up common directory
The .git directory within a worktree is necessarily incomplete.
Typically, it will lack an objects directory, among other things, since
the object storage is shared among all worktrees.  The .git directory
belonging to the main repo contains all these missing items and is
referred to as the common directory.

Add a function to look up the common directory so that we can determine
the storage location for objects.  Explicitly check for the version and
fall back to looking up the plain Git directory if worktree support
isn't available.  This is required since git rev-parse passes through
options it doesn't support, so not checking for the version would mean
that the common directory would always be "--git-common-dir" on older
Git versions.

Update the one existing caller that uses gitobj to use this function for
finding the objects directory correctly.