lfsapi: learn core.sshCommand
We already know how to handle GIT_SSH_COMMAND and its appurtenant shell
parsing requirements.  However, Git also knows the ability to specify an
equivalent command in the config file as core.sshCommand, which allows
for greater flexibility.  Teach git-lfs how to handle core.sshCommand as
well.  Add tests that we process it through the shell correctly and that
GIT_SSH_COMMAND overrides it correctly.

Note that when calling sshParseShellCommand, we specify the default SSH
command directly inline as the fallback command to avoid needing an
additional check for ssh being the empty string in case core.sshCommand
is not set.