t: avoid panic in lfstest-customadapter
The custom adapter test binary wraps the HTTP API client and returns
various responses from the HTTP response to the custom transfer adapter.
This code assumes that if an error occurs, we will always have a valid
HTTP response object, but this is not the case.  If we lack such an
object, we will get a nil pointer dereference, a panic, and a segfault
in the client, causing the test to fail.

While it's currently unknown how to reproduce this issue, it has been
seen on the CI servers.  In order to improve the reliability of the
test, check to see if the response object is non-nil before
dereferencing it and produce a different but distinct error code if it
is nil.