t: ensure client cert integration test runs completely
The shell function that performed the clone for the client certificate
test contained several branches, each of which ended with an exit.  This
caused the test to always exit early (and in the normal case,
successfully) without running the rest of the test.

Change this function to use "return" instead of "exit" so that the test
runs to completion, and fix the single failing assertion.  Ignore this
test on Travis CI, just like we do for the other SSL test, for exactly
the same reason.

Note that this appears to be the only test helper that has this problem:
the other uses of "exit" in the test helpers are all either to skip
tests or to exit unsuccessfully, the latter of which is equivalent to
returning unsuccessfully under set -e (which we use).