subprocess: don't escape shell command due to slash
Add the slash to the list of characters which don't require quoting for
the shell.  No POSIX-compatible shell requires escaping the slash, and
ignoring it for the purposes of escaping lets us write cleaner,
nicer-to-read trace output.

Don't do the same with the backslash, which while being the equivalent
on Windows, is handled specially by POSIX shells.