tq: pass custom transfer adapter args to the shell
When a custom transfer adapter is specified in the configuration, it is
not possible to specify multiple arguments to the process, because only
the last value of the lfs.customtransfer.*.args option is read and the
value is not split.  To make things easier and more flexible,
concatenate the path (after quoting) and the arguments and pass them to
the shell.  Update the documentation to reflect this change.

Update the test custom transfer adapter to parse its arguments (which
are otherwise ignored) and mention them to standard error.  Make the
test check that the arguments are parsed as the shell would expect them
to be.