git/gitattr/attr.go: harden check for assigned attributes
In ParseLines, we read the contents of an io.Reader producing the
contents of a '.gitattributes' file, and fill a slice with `*Line`'s,
one corresponding to each line in the file.

Currently, we handle cases for the following attribute types:

  -foo           set foo to "false"    (unset)
  foo=bar        set foo to "bar"      (set to a value)
  !foo           unspecify foo         (unspecified)

We parse each line by first seeing if it is unset, then if it is
unspecified, and then if it is set to a value.

If we are in the "let's see if it set to a value" and fail to split the
string on the assignment operator, '=', we complain that the attribute
is ill-defined.

Let's harden this check to _only_ assume that it is set to a value when
there is an indicating '=' sign, and temporarily ignore attributes that
don't start with a '-', '!', or contain an '='.