t: add an environment variable to dump test output
When running tests, it's sometimes helpful to see verbose output from a
test even when it succeeds.  This can be useful to compare to a failure
on another system to determine the relevant differences between the
systems.  Add support for reading an environment variable,
LFS_DUMP_TEST_OUTPUT, that if set to a nonempty string causes the
verbose test output to be printed.  Continue printing verbose output
automatically if an assertion fails.