lfsapi: split out persistent credential helpers into a struct
We have several persistent credential helpers which currently live in
the lfsapi client.  In order to make this code reusable among multiple
packages, place all of these credential helpers into a context and
initialize it by using a new constructor.  In addition, store a URL
configuration object, since we'll need that in the context in the
future.