Makefile: build source tarballs for release
Due to changes between gzip versions, the source tarballs produced by
GitHub may change.  To assist distributors, provide stable generated
tarballs as part of the release process.  This allows the distributors
to have a bit-for-bit identical archive that can be verified
cryptographically.

We generate the tarball prefix without the "v" for consistency with
other open-source projects, such as Git, which do not typically include
it.  This means that distributors such as Debian need not repack the
tarball.  We still generate the tarball name with it for consistency
with our other release artifacts, which use it.