script: add a script to update version numbers
The current release process is mostly manual.  To improve automation,
add a script that updates all the various places we store the version
number in preparation for release.  Make the script idempotent.

In order to generate a changelog entry for the Debian changelog, use
the current user's committer identity and a timestamp similar to those
already used.  Specify a time zone of -0000, as that means that the
timestamp is in UTC, but unlike +0000, doesn't imply that we ourselves
are in that timezone.