script: add a noninteractive mode to the changelog generator
While the current changelog generator is useful, it is interactive,
which makes it hard to extract the changelog output from the standard
output of the script.  Provide a noninteractive mode, which specifies
sections but doesn't categorize any of the items.  This can be used
to store the output in a file and allow the user to edit it with the
editor of their choice.  This file can then be programmatically inserted
into the various places the changelog entry is required.