commands/command_migrate.go: fallback to ref.Refspec()
In 'formatRefName()', we special-case references belonging to a remote
in order to round-trip a reference name that 'git-rev-parse(1)' (et.
al.) will understand.

This is done because the `*git.Ref` type does not retain information
about the remote to which it belongs, so we add it back here as a
shortcut to avoid making this change throughout the codebase.

Falling back to ref.Name, however, can introduce ambiguity when there is
a top-level reference of the same name. For example, calling
'formatRefName()' on a reference named 'refs/heads/foo' when there is a
reference named simply, 'foo' will cause Git to complain.

Avoid this by calling `(*git.Ref) Refspec()`, which reintroduces the
prefix thus eliminating any ambiguity.

Co-authored-by: brian m. carlson <bk2204@github.com>