ROADMAP.md: use GitHub issues instead
We introduced the ROADMAP.md file in fd911608 (add a rough roadmap,
2015-04-14), and have updated it 47 times since (roughly ~13 times per
year), with the last update being just under two years ago.

The presence of a ROADMAP.md file is misleading and confusing to users
who don't realize that we use GitHub issues for tracking our new
features.

Since this file is effectively maintained since two years ago (in fact,
many of the items formerly listed in the ROADMAP.md have been completed,
or known to be impossible), let's remove the file entirely.

For readers who are interested in a topic present on the ROADMAP.md,
please search for an open issue that matches your query, since that is
where we are tracking it. If no such open issue exists, and you are
still interested in a feature, please feel more than free to open an
issue and request your feature.