lfsapi: add allowRetry param
Add an allowRetry parameter to DoWithAuth(). This parameter is a
boolean, and allows any requests without an authentication header
to be retried if an authorization error was returned.

This change is required to get things working with the basic
upload adapter. Since it passes along a file reader in the
request body, and since that reader is read by the HTTP Client
in a copy operation to construct a new request, an upload request
body can't be re-used in a new request.