Makefile: trim paths
When building binaries, Go embeds the source filenames into the
binaries, meaning that on most systems, the build directory is embedded
into the binary. This means that binaries are not reproducible when
built by other users and in addition that the layout of the building
user's system is embedded into the binary.

Go provides an option to trim paths out of the ones stored in the
binary, but unfortunately this only works for one path, and since
git-lfs is now built outside the GOPATH to take advantage of the Go
1.11's automatic support for modules, there are two paths that need to
be stripped.

For the benefit of build reproducibility and general build hygiene,
let's strip out the user's home directory when building the binary. This
is better than the status quo, and we can improve this later if
circumstances change. Use "?=" so users who would like to override this
for their particular situation can.

Note that we use "$$HOME" so that the shell expands paths instead of
make, since this means that on Windows, the shell won't ever see any
unescaped backslashes that make might substitute in.