git: introduce constants for index stages
Git can store multiple different versions of a file in the index.
Normally, index stage 0 is the only version provided, but if there's a
conflict, there can be three additional stages.  Since we'll be working
with these stages in a future commit, add constants for them to the git
package.