lfs: add function to create temp files with proper permissions
When we write out a temporary file that we then rename into place, we do
so with 600 permissions, since these are the permissions of the
temporary file.  This can cause problems if the repository needs to be
accessed by a different user, since the umask is not honored.

Add a function to compute the permissions based on the umask and one to
create a temporary file in the repository with the proper permissions.