lfs: honor core.sharedRepository
In order to manage a shared repository easily among multiple users, Git
has the core.sharedRepository variable, which is used to set permissions
for writing objects and files in the repository.  Teach Git LFS to
handle this as well by teaching the existing temporary file code to
honor this setting over the umask.  Since this logic is complex, add
unit and integration tests to ensure that all the code paths are
covered.