locking: write JSON response for "list locks" API call to a cache file
Results of calling "list locks" API will be written in human-readable
format to .git/lfs/cache/locks. Cache files will only be written for
queries which are not limited or filtered. There will be a separate
cache file for every remote ref (Client.RemoteRef).

Cache file format is JSON, which should be considered as implementation
detail and there is no guarantee that the file format will not be changed
for future Git-LFS releases.

The cache file can later be read by Git-LFS or third party tools which wish
to retrieve the last known lock information without invoking
"git lfs locks --cached" (e.g. for performance reasons).

This commit prepares for a fix of issue #3107.