config: add a function for committer timestamp
The author and committer headers in a commit object each contain a name,
email address, time in seconds since the Epoch, and timezone offset.
Since we already have a function to produce the committer name and
email, add one to produce the committer timestamp.

While it may seem bizarre to place timestamp lookup in the configuration
object, it makes sense for two reasons.  First, the configuration object
already stores committer name and email.  Second, in a future commit,
we'll learn about the GIT_COMMITTER_DATE environment variable, which is
best read from the configuration object if present.

Finally, we compute the timestamp at the instantiation of the
configuration object because we plan to also expose author information
and timestamps.  Looking the value up when the function is called could
lead to a race condition where the default author and committer
timestamps differ despite being created by the same process.