CONTRIBUTING.md: use 1-based indices for numbered lists
We use the Markdown trick that the same number over and over again
produces a list of numbers incrementing starting at that base.

This is good, since it reduces the diff later on should the list
ordering change, grow new elements, etc. But it is not desirable to
start the list at zero.

Let's instead increment the list starting at 1, that way we get lists of
the form (1), (2), (3), instead of (0), (1), (2).